requirements/requirements.txt
Django>=3.1.2
django-configurations>=2.3.1
django-extensions>=3.1.5
django-widget-tweaks>=1.4.9
djangorestframework>=3.13.1
isort>=5.10.1
psycopg2>=2.8.6,<2.8.7
requests==2.26.0
drf-spectacular>=0.21.0
